requirements.txt
wiotp-sdk
pyTelegramBotAPI
jproperties
